requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
alembic
psycopg2-binary
python-dotenv
chromadb
openai
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
PyJWT
python-jose[cryptography]
pypdf
requests
pytest
pytest-asyncio
httpx
pytest-cov
werkzeug